mnist

+++ tasks/mnist.py
@@ -37,7 +37,7 @@ class Mnist(Task):
     raise NotImplementedError
 
   def convert_examples(self, examples):
-    return examples['image'], examples['label']
+    return tf.cast(examples['image'], dtype=tf.float32), tf.cast(examples['label'], dtype=tf.int32)
 
   def setup_datasets(self):
     self.info = self.databuilder.info
